+++ Changelog.md
@@ -4,6 +4,7 @@
 - Fix changing/showing tabbar item name for queue (the idea was to hide names for now).
 - Added new view for camera session flow (take a photo -> add details -> take another one...)
 - Added rounding precision (5) for coordinates
+- Upload list is now sorted by `createdDate` if possible (descending)
 
 ## 0.3.0
 - When uploading, we will now disable screen lock timer and enable it back on after the upload finishes/errors out.